# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "get": {
        "operationId": "Properties_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ]
      },
      "post": {
        "operationId": "Properties_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetailDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ]
      }
    },
...
    "/properties/{propertyId}/media/uploads": {
      "post": {
        "operationId": "Media_createUpload",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/properties/{propertyId}/media/{mediaId}/complete": {
      "post": {
        "operationId": "Media_completeUpload",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/properties/{propertyId}/media/order": {
      "patch": {
        "operationId": "Media_reorder",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/properties/{propertyId}/media/{mediaId}/cover": {
      "post": {
        "operationId": "Media_setCover",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ]
      }
    },
    "/properties/{propertyId}/media/{mediaId}": {
      "delete": {
        "operationId": "Media_remove",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ]
      }
    },
...
    "/properties/{propertyId}/rooms/order": {
      "patch": {
        "operationId": "Rooms_reorder",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRoomsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ]
      }
    },
    "/properties/{propertyId}/rooms/{roomId}": {
      "patch": {
        "operationId": "Rooms_update",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "Rooms_remove",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomListDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/properties/{propertyId}/pricing/rooms/{roomId}": {
      "patch": {
        "operationId": "Pricing_updateRoomRates",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomRatesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingDto_Output"
                }
              }
            }
          },
          "default": {
            "description": "Error envelope returned by every failing request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "pricing"
        ]
      }
    },
...
  },
  "info": {
    "title": "d-stay API",
    "description": "Host, admin and guest surfaces for d-stay homestays.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PropertyListDto_Output": {
        "type": "object",
        "properties": {
          "properties": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "DRAFT",
                    "PENDING_REVIEW",
                    "PUBLISHED",
                    "REJECTED",
                    "SUSPENDED"
                  ]
                },
                "membershipRole": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "MANAGER",
                    "STAFF",
                    null
                  ],
                  "nullable": true
                }
              },
              "required": [
                "id",
                "name",
                "status",
                "membershipRole"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "properties"
        ],
        "additionalProperties": false
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
      "PropertyDetailDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_REVIEW",
              "PUBLISHED",
              "REJECTED",
              "SUSPENDED"
            ]
          },
          "membershipRole": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "STAFF",
              null
            ],
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "nullable": true
          },
          "landmark": {
            "type": "string",
            "nullable": true
          },
          "directions": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "district": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "checkInTime": {
            "type": "string",
            "nullable": true
          },
          "checkOutTime": {
            "type": "string",
            "nullable": true
          },
          "houseRules": {
            "type": "string",
            "nullable": true
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "PARKING",
                "HOT_WATER",
                "WIFI",
                "GENERATOR",
                "BONFIRE",
                "PETS_ALLOWED"
              ]
            }
          },
          "mealPlan": {
            "type": "string",
            "enum": [
              "ROOM_ONLY",
              "BREAKFAST",
              "BREAKFAST_DINNER",
              "ALL_MEALS",
              null
            ],
            "nullable": true
          },
          "gstEnabled": {
            "type": "boolean"
          },
          "gstin": {
            "type": "string",
            "nullable": true
          },
          "homestayRegistrationNumber": {
            "type": "string",
            "nullable": true
          },
          "missingFields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "description",
                "photos",
                "rooms",
                "location",
                "city",
                "district",
                "state",
                "checkInTime",
                "checkOutTime",
                "mealPlan",
                "gstin"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "membershipRole",
          "description",
          "rejectionReason",
          "latitude",
          "longitude",
          "landmark",
          "directions",
          "city",
          "district",
          "state",
          "checkInTime",
          "checkOutTime",
          "houseRules",
          "amenities",
          "mealPlan",
          "gstEnabled",
          "gstin",
          "homestayRegistrationNumber",
          "missingFields"
        ],
        "additionalProperties": false
      },
...
      "MediaListDto_Output": {
        "type": "object",
        "properties": {
          "media": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "roomId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "nullable": true
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "PROCESSING",
                    "READY",
                    "FAILED"
                  ]
                },
                "thumbnailUrl": {
                  "type": "string",
                  "format": "uri"
                },
                "cardUrl": {
                  "type": "string",
                  "format": "uri"
                },
                "fullUrl": {
                  "type": "string",
                  "format": "uri"
                },
                "blurhash": {
                  "type": "string",
                  "nullable": true
                },
                "width": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991,
                  "nullable": true
                },
                "height": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991,
                  "nullable": true
                },
                "sortOrder": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "isCover": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "roomId",
                "status",
                "thumbnailUrl",
                "cardUrl",
                "fullUrl",
                "blurhash",
                "width",
                "height",
                "sortOrder",
                "isCover"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "media"
        ],
        "additionalProperties": false
      },
      "CreateUploadDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "byteSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10485760
          }
        },
        "required": [
          "roomId",
          "contentType",
          "byteSize"
        ]
      },
      "UploadIntentDto_Output": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "uploadUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "mediaId",
          "uploadUrl"
        ],
        "additionalProperties": false
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "width",
          "height"
        ]
      },
      "MediaAssetDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "roomId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "READY",
              "FAILED"
            ]
          },
          "thumbnailUrl": {
            "type": "string",
            "format": "uri"
          },
          "cardUrl": {
            "type": "string",
            "format": "uri"
          },
          "fullUrl": {
            "type": "string",
            "format": "uri"
          },
          "blurhash": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "isCover": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "roomId",
          "status",
          "thumbnailUrl",
          "cardUrl",
          "fullUrl",
          "blurhash",
          "width",
          "height",
          "sortOrder",
          "isCover"
        ],
        "additionalProperties": false
      },
      "ReorderMediaDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "nullable": true
          },
          "mediaIds": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "roomId",
          "mediaIds"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR",
                  "PROPERTY_INCOMPLETE",
                  "INVALID_STATUS_TRANSITION",
                  "MEDIA_LIMIT_REACHED",
                  "ROOM_LIMIT_REACHED",
                  "RATE_OVERRIDE_CONFLICT"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "fieldErrors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}